requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-openai>=0.0.5
langgraph>=0.0.20
llama-index>=0.9.48
chromadb>=0.4.22
fastapi>=0.109.2
pydantic>=2.6.1
python-dotenv>=1.0.1
uvicorn>=0.27.1
tqdm>=4.66.1
httpx>=0.26.0
pyyaml>=6.0.1
openai>=1.12.0

# Development and testing dependencies
pytest>=8.0.0
langsmith>=0.0.87
black>=25.1.0
flake8>=7.2.0
isort>=6.0.1
mypy>=1.16.0
pre-commit>=4.2.0
